Give the animations and dwell times a rate
Everything that counted updates now counts time: the laser's temperature and
its reload counter, the info message's dwell, the break pattern's twenty
rings, the game-over hold, the escape-capsule sequence and the ship it
leaves behind, both intro screens, and the explosion cloud.

One new trap, and it is the mirror of the spin peg from the last item. The
capsule's abandoned ship blew up on the tick that *equalled* forty. A whole
tick lands on forty exactly; a third of one steps from just under it to just
over and never lands, and the ship would have flown on intact. It is a
crossing test now, not an equality.

SfxSample counts a sample's life in its caller's ticks, and it is shared
with Stunt Car Racer, which counts in whole updates and says nothing about
time. UpdateSound therefore has two overloads rather than an optional
argument: SCR calls the one that means "a whole tick" and is untouched by
construction, and the test that already covered that path still passes
unchanged, which is what says so.

MessageCount and ExpDelta became floats, so the trace format moves to v3 and
the baselines are regenerated. Every value is unchanged - only the printing
of those two columns - and that was checked against the committed files
rather than assumed. The frame signatures were not regenerated and still
pass, so every pixel is identical too.

CoolLaser moved to a Combat.Laser partial: converting it took the file past
the length the repo enforces, and the two things it counts down belong
together anyway.

+++ src/elite/libs/EliteSharpLib/GameState.cs
@@ -130,7 +130,7 @@ internal sealed class GameState
 
     internal int MCount { get; set; }
 
-    internal int MessageCount { get; set; }
+    internal float MessageCount { get; set; }
 
     internal string MessageString { get; set; } = string.Empty;
